+++ mount/hyperdom.js
@@ -1,0 +1,43 @@
+var Mount = require('./');
+var hyperdom = require('hyperdom');
+var createMonkey = require('../create');
+
+try{
+  var router = require('hyperdom-router');
+} catch(e) {
+  console.warn('no hyperdom router has been found, if you need routing please install this using `npm install hyperdom-router --save-`')
+}
+module.exports = function() {
+  if (router) {
+    router.start();
+  }
+
+  return new Mount({
+    stopApp: function(){
+      if (router) {
+        router.clear();
+      }
+    },
+    startApp: function(){
+      var app = this.app;
+
+      if (Mount.runningInBrowser) {
+        hyperdom.append(Mount.createTestDiv(), app);
+        return createMonkey(document.body);
+      } else {
+        try {
+        var vquery = require('vdom-query');
+        } catch (e) {
+          throw new Error('you must `npm install vdom-query --save-dev` to run tests in node');
+        }
+        var vdom = hyperdom.html('body');
+
+        var monkey = createMonkey(vdom);
+        monkey.set({$: vquery, visibleOnly: false, document: {}});
+
+        hyperdom.appendVDom(vdom, app, { requestRender: setTimeout, window: window });
+        return monkey;
+      }
+    }
+  });
+}
