+++ mount/stubBrowser.js
@@ -1,12 +1,11 @@
 var window = require('global');
 
-
 var registeredEvents = {};
 var pushState, replaceState;
 
 pushState = replaceState = function(state, title, url) {
   window.location.pathname = url;
-  (registeredEvents['onpopstate'] || []).forEach(cb => cb({}));
+  (registeredEvents['onpopstate'] || []).forEach(function(cb) { cb({}); });
 };
 
 window.location = window.location || {};
@@ -14,8 +13,8 @@ window.location.pathname = window.location.pathname || '/';
 window.location.origin = window.location.origin || '';
 window.location.search = window.location.search || '';
 window.history = {
-  pushState,
-  replaceState,
+  pushState: pushState,
+  replaceState: replaceState,
 };
 
 window.addEventListener = function(eventName, cb) {
