+++ package.json
@@ -24,7 +24,6 @@
     "httpism": "^2.6.2",
     "hyperdom": "^0.2.0",
     "hyperdom-router": "^2.18.0",
-    "is-node": "^1.0.2",
     "karma": "1.3.0",
     "karma-browserify": "5.1.0",
     "karma-browserstack-launcher": "1.1.1",
@@ -46,12 +45,18 @@
   },
   "dependencies": {
     "array.prototype.find": "^2.0.3",
+    "browserify-optional": "1.0.0",
     "chai": "3.5.0",
     "debug": "2.6.1",
     "detect-browser": "1.5.0",
     "global": "4.3.0",
     "jquery": "3.1.1",
     "trytryagain": "1.2.0"
+  },
+  "browserify": {
+    "transform": [
+      "browserify-optional"
+    ]
   },
   "keywords": [
     "dom",
